# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Dashboard API",
        "version": "1.0.0",
        "description": "API for Admin Dashboard with Role-Based Access Control",
        "contact": {
            "email": "admin@admin.com"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "UserResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Super Admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@admin.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "super_admin",
                            "admin",
                            "merchant"
                        ],
                        "example": "super_admin"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-08T14:39:44.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-08T14:39:44.000000Z"
                    }
                }
            },
            "MerchantResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "business_name": {
                        "type": "string",
                        "example": "Main Merchant Store"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+6281234567890"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Commerce Street, Jakarta"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-08T14:39:45.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-08T14:39:45.000000Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/merchants": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "summary": "List merchants",
                "description": "Get list of merchants. Super Admin & Admin see all, Merchant sees only own. Supports pagination, sorting, and filtering.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (dropdown: 5, 10, 25, 50, 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [5, 10, 25, 50, 100],
                            "default": 10
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": ["id", "user_id", "business_name", "phone", "created_at"],
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": ["asc", "desc"],
                            "default": "desc"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in business_name or address",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchants retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Merchants retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MerchantResource"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Create merchant",
                "description": "Create a new merchant. Super Admin & Admin only.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "user_id",
                                    "business_name"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "example": "My Store"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+6281234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main Street, Jakarta"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Merchant created successfully"
                    },
                    "403": {
                        "description": "Forbidden - Merchants cannot create"
                    }
                }
            }
        },
        "/api/merchants/{id}": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Get merchant details",
                "description": "Get details of a specific merchant. Role-based access.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant details"
                    },
                    "404": {
                        "description": "Merchant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Update merchant",
                "description": "Update merchant details. Role-based access.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "business_name": {
                                        "type": "string",
                                        "example": "Updated Store Name"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+6281234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "456 New Street, Jakarta"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Merchant updated successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Merchant not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Delete merchant",
                "description": "Soft delete a merchant. Role-based access.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Merchant not found"
                    }
                }
            }
        },
...
    }
}